AI autonomous screening (practical)

This scenario is the same as AI_autonomous_screening_ideal except that the AI reader is not allowed to make a final decision. The no-recall episodes are checked by a simulated fourth reader that is based on the performance of the first reader.

AI_autonomous_screening_practical(
  accession_df,
  reader_df,
  model_pred,
  threshold = 0.05
)

Arguments

accession_df

A data frame; the data frame containing the episodes to be read.

reader_df

A data frame; the data frame containing the reader's read.

model_pred

A data frame; the data frame containing the model prediction.

threshold

A scalar; the threshold for the model prediction.

Value

A data frame; the result of the scenario simulation.