Run simulation analysis

Performs the simulation of the AI replacement scenarios with the second reader arbiter and mixed third reader arbiter respectively.

run_simulation_arbiter_variations(
  accession_df,
  reader_df,
  model_df,
  replacement_threshold,
  bandpass_threshold,
  seed = 1234,
  skip_bandpass = FALSE,
  ...
)

Arguments

accession_df

A data frame; the placeholder for storing the simulation result.

reader_df

A data frame; the data frame containing the reader's read.

model_df

A data frame; the model predictions.

replacement_threshold

The (manufacturer specific) threshold for the replacement scenario.

bandpass_threshold

The (manufacturer specific) threshold for the band-pass scenario.

seed

An integer; the random seed.

skip_bandpass

TRUE / FALSE; whether to skip the bandpass simulation. This is useful during bootstrap. As the bandpass simulation is not stochastic, rerunning the simulation would only generate duplicate results. Hence, this should be set to TRUE for bootstrap.

...

Includes parallel (logical, default to be FALSE), mc.cores (integer, default to be 2), and other optional parameters to be passed to parallel::mclapply.

Value

A list of simulation results